fix ab-tests
